Remove some nulls from exhibitions.js
In the more TypeScript-y parts of the codebase, we assume that place,
promoImage, and format are `T | undefined`, so we use `isNotUndefined`
to filter out missing values.  Because these functions were setting an
empty value to `null`, empty values would sneak past and then cause
errors elsewhere.

For now we explicitly set these to `undefined` instead of `null`, but
longer term getting universal TypeScript and the new fetchers should
fix this.

This fixes an issue with previewing an exhibition; I've tested it
locally.