Use the typed event fetcher, and remember to fetch schedule items
When I did this previously, I called the fetchEvent() function, but I
didn't look at what it was doing -- because getEvent() not only fetches
the event, but also events which are scheduled as part of the event.
Because I was only doing the first call, I missed the second part and
there was missing info on the event page.