Fix the crop of books in their promo
I got confused about the transformPromoToCaptionedImage signature when I converted
it from JavaScript.  The original signature was

    crop: ?Crop = '16:9'

and it was called on the books page with `crop = null`.  If the `crop`
is null, you get the full image.  I thought I could change that to

    crop: Crop | undefined = '16:9'

and then omit the argument on the books page, but I got that wrong --
it caused the book promo to be cropped to 16:9 by default.

Since this function is only used in two places, I've binned the default
and made the use of a 16:9 crop in the one other call explicit.