Split the types for transforming contributors
I noticed this while investigating the contributor link issue -- because
Person and Organisation are slightly divergent types (one uses a string
for the title, the other a structured text field), we lose the benefits
of the type checker here.  In particular, the `sameAs` type was `any`,
but now we know what it is in both branches.