Transform Prismic documents in the API, not the browser
We have a couple of client-side fetchers for content that gets loaded
asynchronously, e.g. the other stories in a series on an article page.

Previously we did this by proxying the Prismic API and transforming it to
our component model in the browser, but this means we have to send all
the Prismic types and transformation logic down to the browser.

By doing the transformation on the server, I'm hoping we can reduce the
bundle size by a non-trivial amount.

I've also cleaned up the client-side fetchers a bit; not all of them are
actually used (or working).