Remove some unused fields in the event transformer
We were getting an error in determineDateRange because calling `reduce()`
on an empty list is an error, but on further inspection, we don't need
this function at all.

Previously the type checker was unable to tell us this, because we called
it in one object then unpacked it into another later:

```typescript
function transformEvent(): Event {
  const event = { dateRange: determineDateRange() }

  // do some stuff

  return { ...event }
```

By consolidating everything into the return statement, the type checker
tell us this field is unused (and several more besides).