Add some tests for formatDateRangeWithMessage
This was a precursor for getting Moment out of this function, but it also
uncovered a bug -- because `isBetween()` does an exclusive range
(i.e. `a < x < b` rather than `a ≤ x ≤ b`), we'd show "Now on" for an
exhibition that was on its final day.  This was caught by the test, and
is now fixed.