Fix a couple of warnings from the type checker
I was getting some warnings about missing return types and use of `any`;
this patch fixes them.  In particular, I've added a `FacilityPromo` type
for use with the component of the same name, which lets me tidy up some
of the facility promo data.

The daily tour promo isn't being used on the what's on page, and the data
is duplicated in the DailyTourPromo component.  We can remove it from the
what's on page and save a bit of data.