Tidy up how we do licenses; toss lodash.merge
We were using lodash.merge to combine two sources of license data.  The combined
data was only being used in the Tasl component, and that component only needed
one source of the license data.  This was completely wasteful and adding an
unnecessary dependency; we can remove lodash.merge and make the license code
easier to follow at the same time.