Don't throw a 500 error on /api/exhibitions-related-content
Normally we pass the exhibition we want to get related content for in
the `params` query parameter, but something has been hitting the API and
not passing that -- then `parsedParams` becomes undefined, and getting
the length of it throws a type error.

You can't do anything useful with this endpoint, so just 400 out.