Don't overwrite the promo field in the Exhibition transformer
The downstream code expects this to be of type `ImagePromo`; a lot of
these fields aren't on ImagePromo, and it's missing the caption which
is used to populate the metadata for social media cards.

We already have the promo from transformGenericContentFields(), so let
that value through rather than replacing it.