Fix a broken link on the homepage
Tracing through the code, the problem is that when we fetch the project,
we don't fetch a `data` field on the document.  This causes transformLink()
to discard the entire link, and then the card doesn't go anywhere.

I don't think we need to check there's data behind the document here --
we only use the ID and type to create the link, so knowing that the link
isn't broken should be sufficient.

(The alternative is updating the fetcher to get the `data` field, but
since we'd instantly discard it that seems wasteful.)