Filter for "interesting" errors in the CloudFront logs
It turns out sending a newline in your URI will cause API Gateway to
return a 503 Bad Gateway error, but there's nothing we can do about that.
Since _somebody_ is doing that by appending data-uri SVGs to their requests,
it's spamming our Slack channel.

Make the bot not do that.