Return a 400, not a 500, for a bad session_token on registration
If we got a session token we couldn't decode on the registration endpoint,
we'd throw an error which got raised as a 500 Internal Server Error.
This is wrong -- we know what the issue is, we got passed bad data by
the user -- so to prevent it showing up in our alerts, raise the correct
error code.