Add a allowRawHtml parameter to <WorkDetailsText>
Initially this is set to true, to mimic the existing behaviour --
but now this choice is configurable and explicit when we use the component.
Currently it looks like we allow raw HTML in *any* field, which probably
isn't what we want.

This opens the door to marking some fields as "should not contain HTML"
in a future patch.