Guard against bad Sierra IDs on the download page
Previously we'd send a download request to DLCS, regardless of what was
passed as the sierraId field.  This could cause 500 errors, if we passed
something that DLCS couldn't handle, e.g.

    /works/amuh57db/download?sierraId=../etc/passwd

By filtering for strings that look like plausible Sierra IDs here, we
reduce the number of obviously bad queries we send to DLCS and cut out
a source of 500 errors.