Fix the appearance of dates in the footer
The problem is that the type checker can't derive a better type for
`data[lowercaseDay][0]` than `any`, so it thinks the `{start,end}DateTime`
values are fine to be passed to `formatTime`, when we need to cast them
to `Date` objects first.

Adding some tests and intermediate types fixes the appearance of opening
times in the footer.