Refactor the way events get grouped in "What's On"
The existing implementation was quite difficult to follow -- it was
broken down into lots of small functions, and it was tricky to see how
they interacted.

It had some quite tricksy behaviour around things like 0-indexed months
(which we don't use anywhere else) and the ordering of events.

This patch replaces it with a hopefully simpler and more readable
implementation:

1. Work out the min/max bounds for the list of events
2. Work out what month that covers
3. For each month, work out (a) what events should be included and
   (b) what order they should appear in

A lot of the intermediate functions are gone, and internally events are
grouped by their English name, not a 0-indexed value.

I've tested it on the current What's On page, which has a somewhat
complex series of events, and it seems to behave correctly.