Simplify uses of classNames that can be plain strings
We only need to use classNames when there's a condition that might mean
a style isn't being used.  If all the conditions are `true`, we can
remove the function and replace it with a string interpolation.

Note: in a few cases, I've made what might seem like non-obvious choices
for formatting the string, e.g. in Quote.tsx:

```
className={
  'quote__cite flex flex--v-end font-pewter' + ' ' + font('intr', 5)
}
```

rather than

```
className={
  `quote__cite flex flex--v-end font-pewter ${font('intr', 5)}`
}
```

This is me fighting the auto-formatter and trying to preserve readability;
in that example, the code gets formatted as

```
className={
  `quote__cite flex flex--v-end font-pewter ${font(
    'intr',
     5
  )}`
}
```

which seems far less readable to me.  I've tried to keep small functions
like `font()` and `grid()` on a single line, even if it means making the
overall className look a bit different.