Link to the raw catalogue API query in the API toolbar
Jonathan and I have just had a confusing incident where we seemed to be
getting mismatched results between the front-end and the API, because
the front-end knew to quote the values, and we didn't.

This adds a link to the catalogue API query from the API toolbar, which
should make this sort of thing easier to reason about -- we can easily
see what queries the front-end is making.