Don't show the quoted query parameters in the filters UI
We add the selected filters inside filterOptionsWithNonAggregates, which
previously took a list of strings -- so a single value would be used for
both the URL query parameter and the user-visible label, which meant it
got quoted.

This patch changes filterOptionsWithNonAggregates to take a list of
`{ label, value }` so these can diverge when necessary.