Make sure the site doesn't crash when we get bad server-data
Last night, the site went down for about a minute.  If you look in the
logs, you see the following error:

    TypeError: Cannot read property 'results' of null
       at simplifyCollectionVenues

This is an error we've seen before, and the previous fix was wrong.

## Context

We fetch certain data from Prismic on a fixed interval (aka "server data").
It gets saved to a JSON file on the server's disk, and read from the file
on every request, rather than going back to Prismic.

If there's an error fetching the data from Prismic (e.g. a timeout), we'd
write `null` into the JSON file.  This causes problems on the next request,
and all subsequent requests until we get new Prismic data.

This gives a plausible explanation for last night's partial outage:

1.  A single instance of the content app has an error fetching server data
    from Prismic.  It writes `null` into the JSON file.

2.  Requests routed to that app start failing.  Other instances have
    valid server data, so they keep serving traffic properly.

3.  A minute later, the instance with bad data gets new server data from
    Prismic.  This succeeds and it writes good data into the JSON file.

4.  All subsequent requests to that app succeed, because now it's got
    valid server data.

This matches the observed behaviour: some requests were failing for about
a minute, then with no intervention the site recovered.

## The previous fix

The previous fix would filter out writing a server data value of `null`,
but this was based on a misunderstanding of the structure of this data.

This is what valid Prismic server data looks like:

    {
      globalAlert: { data: { … } }
      popupDialog: { data: { … } }
      collectionVenues: { results: [ … ] }
    }

and this is what the broken data looks like:

    {
      globalAlert: { data: { … } }
      popupDialog: { data: { … } }
      collectionVenues: null
    }

because the three keys are fetched individually.

The `fetch` method for Prismic server data will *always* return an
object with three keys, because it fetches each of them independently --
and it's the `null` in an object value that's the issue.

## The new fix

There's already code in the function that saves the JSON file to:

* Catch and log an error from the fetcher
* Skip writing to the JSON file
* Schedule the next fetch of the server data

This patch updates the Prismic fetcher to throw an error if it gets bad
data, rather than return a nested `null` value.

Testing this locally by running the content app while offline, I see the
expected error when it tries to write server data:

    Could not update server data key 'prismic'; waiting for next fetch
    interval Error: Failed to fetch globalAlert from Prismic: FetchError:
    request to https://wellcomecollection.cdn.prismic.io/api/v2 failed,
    reason: getaddrinfo ENOTFOUND wellcomecollection.cdn.prismic.io

So I think this will protect us from similar issues in future.