Defer creating the 'Request item' button/modal until we know we'll need it
Running locally on a fairly powerful iMac, this takes the time to expand
the list of 235 items in the original ticket from ~45 seconds to less
than a second -- because they're all manual requests, so we never need to
create this modal component.

It'll still be slow if all the items are orderable online.  Fixing that
feels like it needs a more substantial refactor, possibly to only
create a single instance of that component.