Don't 500 if somebody requests too many pages in an article series
This was a bug that became possible in #7633, when we added pagination
for article series.  The `transformArticleSeries()` method expects a
non-empty list of articles; when we only got the first page, it was
enough to check if the series is empty.  Now we can get higher pages, we
need to additionally check if the selected page is empty.