Group successes on the pa11y dashboard
In the current version of the pa11y dashboard, you have to scroll through
the whole list to check there are no issues – unless you know errors
rise to the top, but that's not obvious.

This groups all the successes at the bottom of the page in a section
titled "No issues reported on the following pages", so when everything
is working it'll be super quick to review the dashboard.