Introduce a WorkBasic type
Previously we'd proxy the entire catalogue API result when showing
a list of works, which is very inefficient -- we use a tiny subset
of the data, and load a lot of works.

In the same style as the Prismic code in the content app, this introduces
a WorkBasic type that just contains the minimal set of fields to show
a search result, and throws away everything else.

<table>
<tr>
<th></th>
<th>Before</th>
<th>After</th>
</tr>
<tr>
<td>/works</td>
<td>541.84 kB</td>
<td>485.31 kB (&minus;10%)</td>
</tr>
<tr>
<td>/concept/v3m7uhy9</td>
<td>223.73 kB</td>
<td>187.92 kB (&minus;16%)</td>
</tr>
</table>

It also continues the gradual adoption of the fetch/transform/types
pattern which we already have for Prismic and which Gareth has been
adding for IIIF.