Don't show exhibitions twice on the homepage
Danny has pointed out that "In Plain Sight" is appearing twice on the
homepage -- once as the top item in the header, once in the list of
current exhibitions and events.

This PR adds a filter so if we're featuring an event/exhibition in the
top spot, it won't appear again in the list.