Include a link to a saved search in Kibana
While debugging with Raphaëlle this morning, she said that her first
point of debugging would be to click the prominent link 'see logs' in
the Slack alert -- even though that takes you to the raw CloudFront log
file, which isn't very useful!

That link was originally meant to help debug the alert, but it's
misleading.  It should be taking you somewhere useful!

This patch improves the alert by adding a link to a canned Kibana logging
search for the front-end apps, which should contain more useful context.
This canned search:

*   pre-filters to the front-end apps
*   adds a query that excludes a lot of logs that aren't interesting for
    debugging, e.g. 200 responses and the health checker

This is the first log in the alert, so people can debug app errors more
quickly.