Pass around day names, not day numbers
The crux of the bug is that Date.getDay() is locale-sensitive, and returns
the day in the user's timezone, not in London.  This can cause issues
when they aren't the same.

This patch replaces it with `formatDayName()`, which works independently
of the user's timezone.  It also allows us to pass around named days of
the week, rather than day numbers, which imo is a bit clearer.