Only fetch an EventBasic for client-side events
We want exactly two fields here: an id and title.  We don't need to
fetch a full event, or more than one; fetching a single EventBasic will
make a (small) reduction in the amount of data here.

Also, I can never remember what example to look at to debug this code,
so add an example link.