Fix a flaky e2e test about scrolling in the item viewer
This test has been failing intermittently (or recently, consistently)
in the desktop e2e tests.  It was failing to find the element which matched
the `data-test-id=active-index` selector.

I had a look at the component, dumped out the HTML, and indeed, that element
may not be on the page if you look immediately after scrolling.
Adding a
short delay seems to make the test pass again.