Improve logging for unexpected errors on search
We can't assume we have an instance of a catalogue API error here, so we
can't use the httpStatus attribute -- if we get a non-API error, this will
return `undefined` and then we get an error from the Node server:

    RangeError [ERR_HTTP_INVALID_STATUS_CODE]: Invalid status code: undefined

This in turn upsets the nginx proxy, with the error:

    upstream prematurely closed connection while reading upstream

This will present to the user as a page which just can't be loaded -- they
won't get our internal error page, their browser will literally get an
empty page.

See https://wellcome.slack.com/archives/CQ720BG02/p1675700908905579

We did see one error from this branch (6 February 2023), when a Prismic query
started returning an Internal Server Error, but the issue resolved itself
before we could investigate properly or reproduce.