Stop sending the Router.pathname value to Segment
This value isn't stable, because it's based on how we structure the
codebase -- it's the path for the current route file [1] -- for example,
if the current route file is `/work.tsx`, then `pathname = '/work'`.

This value can change if we refactor the app internally, even if the
externally-visible URL doesn't change.  To avoid giving the impression
that this is a stable value for analytics, we're going to remove it from
the Segment data.

I've discussed this with Tacey, who's already stopped using it.

[1]: https://nextjs.org/docs/api-reference/next/router#router-object