Bump to the newest version of elastic-apm-node
As of last Friday, GitHub have bumped the default version of Node in
their Ubuntu images to Node 18.  This causes our Actions to fail, because
the version of elastic-apm-node we're running doesn't support Node 18.

This patch bumps to the latest minor version of elastic-apm-node, which
supports Node 18 and should unblock our tests.