Add a note if images are no-tzitzit because copyright
This will make it clearer what's going on, and avoid people thinking the
tzitzit button is toggled off or broken for some other reason.

I think this was a decision made by devs at some point, and tzitzit itself
doesn't support creating attribution strings for in-copyright images --
we may need to tweak this.