Fix the /works/ path patterns in the ALB
When we refactored the routing inside the catalogue app, we didn't update
the path patterns in the ALB rules.  This was causing links from
`/search/works` to `/works` to do a hard reload, dropping the React
SearchContext -- rather than the SPA routing we expect

Adding this path pattern seems to fix the routing, and in particular the
"Back to search results" link on work pages.