Fix our handling of IIIF videos without a "Choice" in the body
This is one of the changes with the new IIIF manifests from DLCS:
previously every video would have a "Choice" in the body, because DLCS
was offering both MP4 and WebM options.

It now only offers MP4, so some videos no longer have a "Choice" -- the
only option is the single MP4.  But we can't remove the "Choice" handling
entirely, because some videos still have a "Choice", e.g. for multiple
resolutions.