Use a vanilla bit of React state in SegmentedControl
This is another case of SegmentedControl (and EventsByMonth) being a bit
of a weird component, and that hiding a bug.

Rather than using some React state with `useState()`, we were passing in
a callback to fire whenever the state value was updated.  Something like:

```typescript
class EventsByMonth extends Component<Props, State> {
  state = {
    activeId: undefined,
  }

  render(): JSX.Element {
    ...
    <SegmentedControl
      id="monthControls"
      activeId={groups[0]?.id}
      onActiveIdChange={activeId => this.setState({ activeId })
    />
  ...
```

Inside SegmentedControl, it was using some regular React state with
`useState`, and we had to call `onActiveIdChange` manually.  The intention
was probably something like:

```typescript
const SegmentedControl = ({ id, activeId }) => {
  ...
  const [activeIdInternal, setActiveIdInternal] = useState(activeId);

  useEffect(() => {
    onActiveIdChange(activeIdInternal);
  }, [activeIdInternal]);
```

but what we were actually doing was:

```typescript
const SegmentedControl = ({ id, activeId }) => {
  ...

  useEffect(() => {
    onActiveIdChange(id);
  }, [activeIdInternal]);
```

This is the correct type, but the wrong value to pass here -- it means
we were calling it with the `id` of the entire component, not the selected
segment.  In the example above we were calling `onActiveIdChange("monthControls")`
instead of `onActiveIdChange("may-2023")` or whatever.

The practical upshot of this is that nothing would appear in a SegmentedControl,
and none of the events are appearing on the What's On page right now.

This `onActiveIdChange` is because we want changes in the SegmentedControl
state to persist into the parent; the better approach is to move that
state into the parent, and pass the getter/setter into SegmentedControl.

So now we have something more like:

```typescript
const EventsByMonth = ({ props }) => {
  ...
  const [activeId, setActiveId] = useState(groups[0]?.id);

  <SegmentedControl
    id="monthControls"
    activeId={activeId}
    setActiveId={setActiveId}
  />
  ...
```

and SegmentedControl is calling the `setActiveId` method from the parent,
and doesn't have any of its own state.

This is more consistent with our other components, and I think easier to
understand.

I had to convert EventsByMonth into a function component so I could use
the `useState` React hook; the actual changes are fairly minor.

(I did try a version of this where I made the bugfix and the style changes
separately, but I introduced a new bug where EventsByMonth was constantly
re-rendering and the page slowed to a crawl, so I went back to this version.)