Remember to mark this <h2> as is-hidden
This got lost in a refactor -- previously it was being added with the
condition `Boolean(activeId)`, which always evaluates to true, but I
misread it as always evaluating to false.

The practical effect is that clicking on the tabs now causes the page
to move around whenever you pick a month on the What's On page.