Fix a React warning about duplicate keys in the breadcrumb
On the Press page, there's a warning about duplicate keys, because the
breadcrumb goes:

    Home | About us | (Part of) About us

and so we have two breadcrumb entries with a key `About us`.

Arguably this points to a data issue, but in the meantime we can make
React happy by including the prefix in this key to distinguish them.