Ensure we pass iiifImageLocation into IIIFViewer
In 2d7612e we removed the `iiifImageLocation` prop, because we can derive
it from the Image -- but we can only do this on the /images page.

When we're on the /items page, the Image is undefined so we can't get
a IIIF image location inside the IIIFViewer component.  This means that
anything which is a single image doesn't appear on the items page.

Passing the `iiifImageLocation` explicitly seems to fix these broken
images; we should add a test to catch this, but I don't want to block
fixing these pages behind that change.