Get PDFs from the new-style manifests from DLCS
See the comment in getPdf() --  the way we can find PDFs in the manifests
has changed slightly, so now we support both old and new methods.

The new manifests include a label for the rendering which isn't useful --
just `{ 'none': ['-'] }`, which I've changed to map to `undefined` rather
than spitting out a hyphen which isn't useful.