Reduce the amount of data we fetch on item pages
While testing some stuff for the new IIIF manifests, I noticed that we're
loading a lot of work fields for the item page, even though only a handful
of them are used.  These fields can be quite large, e.g. the page I'm
testing has a large archive tree which is completely wasted.

Only fetching the fields we need should reduce the size of these pages.