An event is past as soon as it's done, not the next day
This changes the behaviour of `isEventPast` slightly, to avoid confusion
on the same day as an event but after the event has finished.

This function is used on event schedule items to determine whether to
show booking information below a scheduled event, e.g. "Just turn up"
or a link to Eventbrite tickets.  But it wouldn't mark an event as past
until the end of that day, meaning an event could still be showing
"Just turn up" even after it's over!

e.g. an event finishes at 2pm, and somebody looks at the site at 3pm.
Because `isEventPast` looks to see if the day of the event has passed,
it says "false" and the page shows "Just turn up".  This is misleading!
There's nothing to turn up to!

Now it marks events as past as soon as the event ends, avoiding the
possibility of confusion.

(It also lets us ditch another use of UTC date comparisons!)