Add the concurrency group to _actually_ running the e2es
In my previous change, I added a concurrency group to the e2e trigger --
so the e2es wouldn't be triggered at the same time as a running deploy,
but they could run at the same time as a deploy.  I think adding the
concurrency group here should fix that.