An event is upcoming if it hasn't finished yet
The naming here is a bit misleading -- we want upcoming or in-progress
events to show under the 'Events' header, because if an event isn't
upcoming it moves straight to 'Past events'.

This is causing issues for the Land Bodies Ecology Festival, which is
already showing Day 1 under 'Past events' even though it's still
happening today.