Simplify the data we pass to concept pages
Currently we pass concept pages the complete CatalogueResultList type,
which includes a bunch of data about pagination that we don't need --
we only care about the list of works/images, and how many there are in
total.

This is the same as the search overview page, which already has some
types and helpers for dealing with this sort of result -- we can reuse
them here.