Don't include the description field on WorkBasic
Some of the works have very large descriptions which we send to search
results, even though they're never used -- we only want them in the IIIF
viewer (and even then I wonder if they're useful).

Using `/search/works?query=copy%20url` as an example:

* Before = 389.93 kB
* After  =  18.35 kB

That's a ~95% saving in the size of the Next.js props.