Use named variables and helper functions in opening-times tests
This is a straight-up readability improvement: we have a bunch of values
in these tests which feature a lot of repetition and boilerplate.

This patch tries to make the tests more readable in two ways:

*   Hiding some of the boilerplate behind named functions, so there's
    less repetition

*   Using named variables to make it more obvious when the same value
    appears in multiple places.  This should make it easier to see what
    the test is doing also -- you can read a list of variable names
    rather than JavaScript objects.

This patch only changes the tests, and only changes variable names -- none
of the values or "live" code have been changed.