We can collapse these two lines into a single loop
Here the `groupedExceptionalDays` variable is a list of lists:

    [
      [Date1, Date2, Date3]
      [Date4, Date5]
      [Date6, Date7, Date8, Date9]
    ]

We're looking for a single date in this list which matches `isSameDay`.

The first line finds the group/inner list which contains a matching date,
then the second line finds the matching date within that list.  By
flattening the dates into a single list, we make this loop a bit simpler
and a bit faster.