Convert another use of isSameDay to use London comparisons, not UTC
In this function, we're building up a list of all the dates when we have
exceptional opening hours, e.g. if we had two venues

    library.exceptionalOpeningHours = [3 Jan, 2 Jan, 1 Jan]
    gallery.exceptionalOpeningHours = [3 Jan, 4 Jan, 5 Feb]

then this would return

    [1 Jan, 2 Jan, 3 Jan, 4 Jan, 5 Feb]

I expect these dates to be consistent for each batch that's entered
(e.g. if somebody enters five dates, they'll have the same UTC offset),
so we're comparing like-for-like in this function.

This means that if somebody enters the same override date in library and
gallery, I expect it would be an _identical_ date, not merely a different
time on the same day.

In practice this means we could actually replace this with the even
stronger check (ignoring any weirdness around comparing dates in JS):

```javascript
const isAlreadyInResult = result.some(t =>
  t.overrideDate === thisOverride.overrideDate
);
```

and we'd get the same results.

We only care about them being the same day, but for this check I don't
think the London/UTC distinction will make any difference and so we can
switch to the default of London-based comparisons.