Use London comparisons in groupConsecutiveExceptionalDays, not UTC
As in a5eac3d, I think we could actually replace this with the stronger
assertion:

```javascript
addDays(date.overrideDate, -1) === lastDayOfGroup
```

because any time these days are consecutive, they'll have been entered
as part of the same batch and I'd expect them to have the same time of
day.  As such, within this function I'm pretty sure UTC and London will
always behave in the same way.

(And all the tests pass unmodified.)