We can use UTC comparisons here, not London
In this part of the code, we've got a complete list of all the days in
the exceptional opening period, and we're trying to work out which
opening hours to use in that time.

e.g. if we have exceptional hours on Mon/Weds/Fri, then this function is
looking at Mon/Tues/Wed/Thurs/Fri:

* Mon = use the exceptional hours
* Tues = use the normal hours
* Wed = use the exceptional hours
* Thurs = use the normal hours
* Fri = use the exceptional hours

What's happening in this code is we're going through the days in this
period one-by-one, and looking to find if there are exceptional hours
for this day.  As with previous PRs, it should be sufficient to do

```javascript
day.overrideDate === date
```

here because all the dates come from Prismic in the same format.

Using `isSameDay` is a bit safer, but there shouldn't be any difference
between UTC/London in practice.